--- docs/source/plugins/chatgpt.rst
@@ -1,12 +1,0 @@
-.. _chatgpt:
-
-#####################
-ChatGPT API reference
-#####################
-
-.. tags:: Integration, OpenAI
-
-.. automodule:: flytekitplugins.chatgpt
-   :no-members:
-   :no-inherited-members:
-   :no-special-members:
