+++ docs/source/plugins/openai.rst
@@ -1,0 +1,12 @@
+.. _openai:
+
+################
+OpenAI reference
+################
+
+.. tags:: Integration, OpenAI
+
+.. automodule:: flytekitplugins.openai
+   :no-members:
+   :no-inherited-members:
+   :no-special-members:
