--- docs/source/plugins/openai_batch_api.rst
@@ -1,12 +1,0 @@
-.. _openai_batch_api:
-
-##########################
-OpenAI Batch API reference
-##########################
-
-.. tags:: Integration, OpenAI
-
-.. automodule:: flytekitplugins.openai_batch_api
-   :no-members:
-   :no-inherited-members:
-   :no-special-members:
