--- plugins/flytekit-openai/flytekitplugins/chatgpt/__init__.py
@@ -1,13 +1,0 @@
-"""
-.. currentmodule:: flytekitplugins.chatgpt
-
-.. autosummary::
-   :template: custom.rst
-   :toctree: generated/
-
-   ChatGPTAgent
-   ChatGPTTask
-"""
-
-from .agent import ChatGPTAgent
-from .task import ChatGPTTask
