+++ plugins/flytekit-openai/flytekitplugins/openai/__init__.py
@@ -1,0 +1,21 @@
+"""
+.. currentmodule:: flytekitplugins.openai
+
+.. autosummary::
+   :template: custom.rst
+   :toctree: generated/
+
+   ChatGPTAgent
+   ChatGPTTask
+   BatchEndpointAgent
+   BatchEndpointTask
+   download_files
+   upload_jsonl_file
+   create_batch
+"""
+
+from .batch_api.agent import BatchEndpointAgent
+from .batch_api.task import BatchEndpointTask, download_files, upload_jsonl_file
+from .batch_api.workflow import create_batch
+from .chatgpt.agent import ChatGPTAgent
+from .chatgpt.task import ChatGPTTask
