--- plugins/flytekit-openai/flytekitplugins/openai_batch_api/__init__.py
@@ -1,17 +1,0 @@
-"""
-.. currentmodule:: flytekitplugins.openai_batch_api
-
-.. autosummary::
-   :template: custom.rst
-   :toctree: generated/
-
-   BatchEndpointAgent
-   BatchEndpointTask
-   download_files
-   upload_jsonl_file
-   create_openai_batch
-"""
-
-from .agent import BatchEndpointAgent
-from .task import BatchEndpointTask, download_files, upload_jsonl_file
-from .workflow import create_openai_batch
