+++ plugins/flytekit-openai/tests/openai_batch/test_workflow.py
@@ -1,8 +1,8 @@
-from flytekitplugins.openai_batch_api import create_openai_batch
+from flytekitplugins.openai import create_batch
 
 
 def test_openai_batch_wf():
-    openai_batch_wf = create_openai_batch(
+    openai_batch_wf = create_batch(
         name="gpt-3.5-turbo",
         openai_organization="testorg",
     )
